Fixed street bets
-Fixed street bets by adding a hidden button as it did not work on android (not a great fix but couldn't work out any other way)

- Reworked the reset coins script so now its used if the balandamount file is empty (first time install, will remove when apart of AGH)

+++ AGH-Roulette/Assets/Scripts/Amount Bet Scripts/ResetCoins.cs
@@ -6,9 +6,12 @@ public class ResetCoins : MonoBehaviour
     public void ResetFile()
     {
         string path = "/balandamount.txt";
-        StreamWriter writer = new StreamWriter(Application.persistentDataPath + path);
-        writer.Flush();
-        writer.WriteLine("Coins: 500");
-        writer.Close();
+
+        if (!File.Exists(Application.persistentDataPath + path))
+        {
+            StreamWriter writer = new StreamWriter(Application.persistentDataPath + path);
+            writer.WriteLine("Coins: 500");
+            writer.Close();
+        }
     }
 }
